Move components into their own directory for better sharing
